pyproject.toml
[tool.poetry]
name = "rag-article-search"
version = "0.1.0"
description = "An advanced Retrieval-Augmented Generation (RAG) system."
authors = ["John He <johnhe@u.northwestern.edu>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
torch = ">=2.5.1"
python-dotenv = ">=1.0.1"
pyyaml = ">=6.0.1"
kaggle = ">=1.6.14"
kagglehub = ">=0.2.0"
pandas = ">=2.2.2"
tqdm = ">=4.66.4"
langchain = ">=0.2.5"
langchain-text-splitters = ">=0.2.1"
langchain-huggingface = ">=0.0.3"
langchain-pinecone = ">=0.1.1"
langchain-core = ">=0.2.9"
pinecone = ">=5.0.0"
langgraph = ">=0.1.0"
google-generativeai = ">=0.8.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.0"
pytest = "^8.4.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"